refactor(vla): clean up LingBot VLA v2 integration
Split attention and alignment helpers out of the checkpoint loader, remove unused tokenizer and loading code, and scope CUDA runtime configuration to pipeline construction.

Add Transformers 5.14 compatibility for the custom Qwen layers and expose CUDA Graph configuration through direct inference and the native service.

Verification:
- 73 focused tests passed with Transformers 5.14.1
- 73 focused tests passed with Transformers 4.57.3
- H100 BF16 frozen-input regression passed
- H100 fused-fp8-graph CUDA Graph regression passed
- ruff check, ruff format --check, and git diff --check passed
